refac send request test
